requirements/prod.txt
uvloop==0.17.0; os_name != 'nt'
hikari[speedups]==2.0.0.dev116
hikari-crescent[cron]==0.5.2
hikari-flare==0.1.0b4
pyyaml==6.0
asyncpg==0.27.0
attrs==22.2.0
